+++ plugins/basestate/include/yarp/dev/BaseStateDriver.h
@@ -14,16 +14,10 @@
 #include <gazebo/physics/Link.hh>
 #include <gazebo/physics/physics.hh>
 
-#if GAZEBO_MAJOR_VERSION >= 8
 #include <ignition/math/Pose3.hh>
 #include <ignition/math/Vector3.hh>
 #include <ignition/math/Matrix3.hh>
 #include <ignition/math.hh>
-#else 
-#include <gazebo/math/Pose.hh>
-#include <gazebo/math/Vector3.hh>
-#include <gazebo/math/Matrix3.hh>
-#endif
 
 #include <yarp/dev/DeviceDriver.h>
 #include <yarp/dev/IAnalogSensor.h>