+++ plugins/multicamera/src/MultiCameraDriver.cpp
@@ -328,7 +328,6 @@ bool yarp::dev::GazeboYarpMultiCameraDriver::getRgbIntrinsicParam(yarp::os::Prop
         Distortion* distModel = camPtr->LensDistortion().get();
         if(distModel)
         {
-#if GAZEBO_MAJOR_VERSION >= 8
             intrinsic.put("k1",              distModel->K1());
             intrinsic.put("k2",              distModel->K2());
             intrinsic.put("k3",              distModel->K3());
@@ -336,15 +335,6 @@ bool yarp::dev::GazeboYarpMultiCameraDriver::getRgbIntrinsicParam(yarp::os::Prop
             intrinsic.put("t2",              distModel->P2());
             intrinsic.put("principalPointX", distModel->Center().X());
             intrinsic.put("principalPointY", distModel->Center().Y());
-#else
-            intrinsic.put("k1",              distModel->GetK1());
-            intrinsic.put("k2",              distModel->GetK2());
-            intrinsic.put("k3",              distModel->GetK3());
-            intrinsic.put("t1",              distModel->GetP1());
-            intrinsic.put("t2",              distModel->GetP2());
-            intrinsic.put("principalPointX", distModel->GetCenter().x);
-            intrinsic.put("principalPointY", distModel->GetCenter().y);
-#endif
         }
         else
         {